pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "elm-data-pipeline"
version = "0.1.0"
description = "Data preparation pipeline for ELM using WikiText-103 and Qwen3-Embedding-4B"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.51.0",
    "datasets>=2.0.0",
    "polars>=0.20.0",
    "safetensors>=0.4.0",
    "tqdm>=4.65.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "jsonlines>=3.0.0",
    "faiss-cpu>=1.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[tool.setuptools]
packages = ["Data_Preparation.data_pipeline", "Data_Synthesis.synthesis_pipeline"]